Basice implementation for user registration

+++ authentication 1/src/main/java/com/csmarton/auth/repository/UserRepository.java
@@ -2,8 +2,9 @@ package com.csmarton.auth.repository;
 
 import com.csmarton.auth.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 }
